refactor report package seams
